Use MapContainer to filter maps to be cleaned up when migrating off a partition [HZ-2643] (#24844)
When migrating a partition off from a member,
`PartitionContainer#cleanUpOnMigration` can recreate an already removed
`RecordStore` instance while filtering the maps that have to clean up
based on their backup count. It is doing so only to get access to the
backup count of the maps defined in their config. These changes make
`cleanUpOnMigration` accessing the config directly through the
`MapContainer`, with which the `RecordStore` is not needed.