Implement member pruning based on required partitions [HZ-1659] (#24689)
This patch is introducing **member pruning** -- 
a beneficial optimization technique for small jobs.
The overwhelming majority of such jobs are generated by SQL, 
and our team effort is to replace Predicate API,
where `PartitionPredicate` is available for partitioned data querying. 
For that, we can use SQL optimization phase also for determining partitions 
and for all relations, we can extract information about required partitions to run to `JobConfig`. 
This information in `JobConfig` is an indicator for `ExecutionPlanBuilder` to pick 
cluster members which are required for the job.