Fix WAN events fired after split-brain merges where values don't change
The current mechanics of WAN replication mean that `MergeOperation` (for
`IMap) and `CacheMergeOperation` (for `ICache`) invocations generate a
WAN replication event for all results besides no data present on both
sides of the merge. This means that a scenario where the values of the
merge are identical, and only expiry metadata is updated, still causes
WAN replication to take place.

This is not desired behaviour as it may introduce unexpected results on
the WAN replication target, since the source cluster has not actually
had any data changed, it simply recovered from a split-brain scenario.
One example of this would be where a target cluster processes WAN
replicated data from a source cluster, then deletes it from its local
data set once processed. A split-brain merge on the source cluster would
cause data that has already been processed on the target cluster to be
re-introduced for unnecessary processing.

This commit aims to resolve this situation by altering the returned type
of the `RecordStore#merge` and `AbstractCacheRecordStore#merge` methods
to provide a more verbose result for the merge operation; this allows us
to maintain all of the existing functionality, while also allowing the
WAN replication stage in `MergeOperation` and `CacheMergeOperation` to
check this more verbose response and avoid sending a WAN Replication
event when the response is `VALUES_ARE_EQUAL`. The key to this change is
that it only changes WAN replication behaviour in this circumstance, and
all existing functionality (such as calling entry listeners) remains.

I introduced separate response classes for `IMap` vs `ICache` as their
functionality and results offered differ slightly; we could get away with
using a shared response class and adjusting the internals, but to allow
for easy expansion and possible divergence in functionality between the
two data structures in the future, I opted to use separate classes.

This solution does break the existing API for the `#merge` function in
`RecordStore` and `ICacheRecordStore`, but both classes are within our
`impl` packages, and so we make no API guarantees about them.