Retry in KinesisSinkP#saveToSnapshot [HZ-2545] [5.2.z]
When `saveToSnapshot()` is called while `sendResult` is non-null then
the result might contain failed records which are retained in the
buffer.

However the buffer is not checked if it is empty, which leads to losing
the items in the buffer.

The behaviour of `saveToSnapshot` should be the same as `complete()` -
it needs to flush all records and retry on any failures. The change
simply delegates to `complete()`.

Fixes #24774
Fixes #23692