Retry in KinesisSinkP#saveToSnapshot [HZ-2545] (#24779)
When `saveToSnapshot()` is called while `sendResult` is non-null then
the result might contain failed records that are retained in the buffer.

However, the buffer is not checked if it is empty, which leads to losing
the items in the buffer.

The behavior of `saveToSnapshot` should be the same as `complete()` - it
needs to flush all records and retry on any failures. The change simply
delegates to `complete()`.

Also updated Localstack to 2.1.0, which supports arm. The previous
issues with never image seem to be fixed.

Fixes #24774
Fixes #23692