Don't fail consistency check when partitions not inited
The consistency check in DataConnectionConsistencyChecker needs to
iterate over sqlCatalog's values. When partition assignments are not
done and the cluster state does not allow triggering partition
assignments (for example in FROZEN state), the consistency check should
be skipped, because no data can exist in the sqlCatalog anyway.
Previously, the check would fail with an IllegalStateException.

Fixes HZ-2716