Catch STARTING -> RUNNING transition in JobStatusListenerTest [HZ-2447] (#24659)
The tests in JobStatusListenerTest wait for listener registration to "run." In this way, only the first two status transitions, namely NOT_RUNNING → STARTING and STARTING → RUNNING, might be missed and the tests tolerate this by performing tail comparison. testListenerDeregistration(), however, only checks the first two transitions, which causes it to proceed when no event is received at all. As a solution, the TestSource is modified to wait in the "init" phase, so that STARTING → RUNNING transition is always caught.

EE PR: hazelcast/hazelcast-enterprise#6209
Fixes hazelcast/hazelcast-enterprise#6008