Socket scheduling improvements
There were problems with many connections because dirty sockets
get processed as a task on the reactor and this leads to interleaving
writing to the socket with ready events on the sockets and NIO doesn't
like that. So sockets are no longer scheduled directly as task on
the socket. This also gives us the ability to do better socket
scheduling in the future.

Also some API cleanup that got triggered based on the socket scheduling
cleanup.