Improve hit/miss cache statistics counter performance
In #21676 it was raised that in some scenarios, the statistic collection
process was introducing significant overhead on NearCache performance.

Specifically, a hit / miss counter is implemented using a
`AtomicLongFieldUpdater` - changing this to a `LongAccumulator` improves
performance.

Depending on the benchmark, the improvement was between 2% & 60% -
consistently faster, but obviously workload dependent.

Fixes [#21676](https://github.com/hazelcast/hazelcast/issues/21676) /
[HZ-2821](https://hazelcast.atlassian.net/browse/HZ-2821)