Only skip broadcast for joining members, cleanup test
Skipping broadcast to all members entirely causes regression in sequential joins - so the best we can do is skip broadcasting to members in the join batch.

This commit also removes ClusterBatchedJoinTest in favour of it replacing the current contents of EventRegistrationTest; this new test covers more scenarios than both previous tests individually.