Move PartitionMigrationComputeIntensiveTest to NightlyTest [HZ-2623]
As the name suggests, this test is very computationally expensive, and
as such is prone to false flag failures on build executors that are
already under pressure during tests (such as Sonar builds). This is
shown in https://github.com/hazelcast/hazelcast/issues/24791 where the
recently added logging shows all nodes experience slow operations while
warming up partitions, and it is simply a case of all processing being
hampered; this is demonstrated by logs showing the slow operations do
make progress, but not fast enough, while health monitor shows system
load at maximum capacity.

Moving this test to `@NightlyTest` will stop it being run on Sonar
builds and keep it reserved to the nightly executions where pressure
is reduced. Another approach would be to reduce the `PARTITION_COUNT`
for this test, but as this test is designed to be intensive, I think
that should be kept as a last resort.