Create ClientProperty for PARTITIONING_STRATEGY_CLASS and use on clients [HZ-2879] (#25193)
There is the `HazelcastProperty` of `PARTITIONING_STRATEGY_CLASS` (which
has property key `hazelcast.partitioning.strategy.class`), which is used
on both the member-side and client-side. Currently, we are referencing
the `ClusterProperty` class for this from the client, but really this
property should be contained within `ClientProperty` for consistency.

I have copied the property from `ClusterProperty` into `ClientProperty`,
maintaining the existing property name as it has been the same for the
last 8 years, and we do not want to break backwards compatibility. By
including this property in the `ClientProperty` class we make it clear
to everyone (including the documentation team) that this property is
valid for use on the client as well.

Fixes https://hazelcast.atlassian.net/browse/HZ-2879