Cancel read from Ringbuffer reliable topic listener is removed [hz-2930]
In the current design the listener calls ringbuffer.readManyAsync() and then waits for at least one message. Is no message is sent, the listener is not GC'ed.
If a message is sent, then readManyAsync() completes and listener is removed. So effectively the listener gets removed with a lag. One received message later
We can fix it by cancelling the Future for readManyAsync() when listener is removed.