Cancel read from Ringbuffer reliable topic listener is removed [hz-2930] (#25246)
In the current design the listener calls ringbuffer.readManyAsync() and
then waits for at least one message. Is no message is sent, the listener
is not GC'ed. If a message is sent, then readManyAsync() completes and
listener is removed. So effectively the listener gets removed with a
lag. One received message later. We can fix it by cancelling the Future
for readManyAsync() when listener is removed.

Fixes:  https://github.com/hazelcast/hazelcast/issues/15409
Jira : https://hazelcast.atlassian.net/browse/HZ-2930

Breaking changes (list specific methods/types/messages):
* API
* client protocol format
* serialized form
* snapshot format

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Label `Add to Release Notes` or `Not Release Notes content` set
- [x] Request reviewers if possible