AsyncFile IntPromise simplification.
The old design whereby you get an IntPromise
from the allocator was too slow; especially with
the Nop benchmark this became visible.

With the new design instead of returning an IntPromise,
a callback is passed.

The IntPromise implements this callback, so you can pass an
IntPromise as well.

And because the  IntPromise is owned by the caller, the caller
can decide if it wants to pool the IntPromise or can store it
locally a lot cheaper without going through the allocate/free
overhead of the IntPromise.