UringFifoStorageAllocator allocation perf improvement
begin allocating from the beginning of the array.

The JVM will always need to load the beginning of the array anyway, so when
you allocate, it is better to allocate from there since that will be in the
cache. If you begin from the end, you need to load both the beginning and
end in the cache.