Do not throw and log NPE and finish operation logic when clearing empty IMap (#25202)
Executing `IMap.clear()` (eg. `instance().getMap("whatever").clear()`)
on non-existent IMap caused the following exception to be logged:

```
13:21:33,052 ERROR || - [ClearOperation] hz.SqlFilterProjectTest_vibrant_tharp.partition-operation.thread-1 - [127.0.0.1]:5701 [dev] [5.4.0-SNAPSHOT] Cannot invoke "com.hazelcast.map.impl.MapContainer.hasInvalidationListener()" because "this.mapContainer" is null
java.lang.NullPointerException: Cannot invoke "com.hazelcast.map.impl.MapContainer.hasInvalidationListener()" because "this.mapContainer" is null
	at com.hazelcast.map.impl.operation.MapOperation.invalidateAllKeysInNearCaches(MapOperation.java:406) ~[classes/:?]
	at com.hazelcast.map.impl.operation.ClearOperation.afterRunInternal(ClearOperation.java:73) ~[classes/:?]
	at com.hazelcast.map.impl.operation.MapOperation.afterRun(MapOperation.java:269) ~[classes/:?]
	at com.hazelcast.spi.impl.operationservice.impl.OperationRunnerImpl.afterRun(OperationRunnerImpl.java:384) [classes/:?]
```

The exception was logged on ERROR level and not propagated to caller.
Some parts of `afterRun` were not executed. I am not sure how important
are they in this case.

This error is possible in Jet which tries to clear snapshot IMaps before
writing to them. On first use the IMap does not yet exist.