Restore implements Versioned where required
Cleanup of rolling-upgrade code paths in #24595 is the cause
of compatibility regressions due to eagerly removing Versioned
interface. Proper code cleanup for rolling-upgrade conditional
serde requires that the Versioned interface is still implemented
one version after the conditional serde code is removed, to ensure
that major.minor version bytes are still written to the network
output and previous-version serde will find the proper in.getVersion
to deserialize correctly.