AsyncServerSocket bind simplification.
Instead of having a bind/listen function on the AsyncServerSocket,
it is now done through the AsyncServerSocket.Builder.

This simplifies the state of the AsyncServerSocket since it always
starts being with listen(fd..) and already bound.

Also makes the API simplier since there is less to call.

Another advantage is that the localAddress/port can be cached
and there is no further object creation or syscalls involved.