Tpc engine advanced socket createdestroy tests (#25311)
A redesign of how socket closing works with io_uring.

So instead of just closing the socket, first a shutdown is done and then
a close. If you start with a close and there is a pending e.g. read, the
read will not be cancelled and no FIN/RST will be send, so the other
side isn't notified.

Also there was a lot of change with the completion queue handler design
because you need to be very careful with deregistering.