Lot of work on DeadlineScheduler.
Moved all the methods for scheduling to the DeadlineScheduler.

DeadlineScheduler pools DeadlineTasks.

Improved testing DeadlineScheduler; but more work is required.

Added a soak test for the deadline scheduler to ensure that
running for a long period with a large number of deadline tasks
isn't leading to problems.

The DeadlineScheduler doesn't return promises, but uses a callback.

This callback is implemented by the promise, so that the caller can
decide if it wants to use a promise or not and in case of a promise,
if it wants to pool or not. Similar to the design of the storage
scheduler.