GitHub New Description
@fawazghali has requested the proposed changes in this PR for the "What is Hazelcast" section of the README.md file. 
His another request is to update the "About" section of the hazelcast/hazelcast repo as follows: "Hazelcast is a unified real-time data platform combining stream processing with a fast data store, allowing customers to act instantly on data-in-motion for real-time insights." I can not send a PR (it is a manual field update) for changing "About" section of the repo. I'll assume your approval to this PR is also an approval the proposed change for the "About" section of the repo.