TaskRunner can now transform task.
When the TaskRunner received e.g. an IOBuffer, it can transform the
task to some Task implementation. The problem is that it had no way
telling the TaskQueue about the transformed version.

That is why the TaskRunner now receives a task reference that can
be updated by the TaskRunner and the TaskQueue will put whatever
the value is on the task reference back on the task queue when
needed.