Introduced Signals abstraction.
With the Signals there is a central mechanism where 'actions' from outside
of the reactor can be made visible to the reactor to get them processed.

The 2 main sources of such actions are TaskQueues and AsyncSocket (for writing).

A problem in the previous design was blocking 'concurrent' taskQueues because
their scheduling complexity is O(N) EVEN when the taskQueues are blocked. This
is because they formed a linked list and needed to be scanned frequently for
any pending work.