[TPC] Introducing Signals (#25448)
The way outside threads notify the reactor of pending work has been
cleaned up.

There were 2 mechanisms:
1) network scheduler has a mpsc queue for dirty sockets
2) concurrent task queues were stored in a linked list.

Now there is a single mechanism called 'Signals' which contains a
mpsc-queue which is guaranteed to be checked periodically. And if there
is a task queue with work outside the reactor or socket that is written
to outside of the reactor, they add a signal action to that queue which
moves the socket/task queue to the appropriate internal task queue for
further processing.

Since task queues can be concurrent, the creation of the task queue has
been moved from the eventloop to the reactor because that is where other
concurrent resources like sockets can be created.

The spin flag has been replaced by an IdleStrategy (from Agrona).