Throw user cancellation exception if the job was cancelled before it was initialized [5.3.z] (#25452)
Wrong exception was thrown if the job was cancelled very early by the
user. Instead of specific `CancellationByUserException` the generic
`CancellationException` was thrown. This caused some flakiness in unit
tests (see eg.
https://jenkins.hazelcast.com/job/Hazelcast-pr-builder/20340/testReport/junit/com.hazelcast.jet.impl/JobSummaryTest/when_batchJob_cancelled/)

Not adding unit test because there does not seem to be an easy way to
trigger this situation.

Backport of: #25383