Reproduce duplicated index scan results with conjunction of overlapping criteria [HZ-3012], [HZ-3013], [HZ-3014] (#25254)
SQL queries can produce duplicate results when conjunction of
overlapping criteria is used with index scan. Currently we support index
scan with conjunctions (which maps to `IndexComponentFilter`) only for
equality predicates but it is possible to trigger the problem by having
query with `WHERE indexed_column=? or indexed_column=?` and passing the
same parameter value. There are also other issues with correctness of
the results, especially in `ORDER BY ... DESC` queries.

This PR adds private property `hazelcast.sql.index.hint.fetch.size` which
is used in tests. It could be also used to workaround some of the
problems (like HZ-2881) if increased, but at the cost of increased
memory consumptions or OOM in extreme cases.

Reproducer for HZ-3012, HZ-3013, HZ-3014
Tests reproducing problems are disabled.