Use tmpfs for MySQL database provider [HZ-2430] [5.3.z] (#25468)
The test inserts 10k items in 10 iterations. This results in 10k
transactions, and MySQL in a Docker container is slow when committing
the transactions.

This commit uses tmpfs for the database filesystem, effectively running
it in memory. Provides roughly 10x speedup on my local machine.

Fixes #24549

Backport of https://github.com/hazelcast/hazelcast/pull/25467