Print a warning when British spelling of "License" is used (#25460)
[Some users](https://github.com/hazelcast/hazelcast-docker/issues/68)
have been struggling with the `hazelcast.enterprise.license.key`
property, instead using the `hazelcast.enterprise.licence.key` property,
which is silently ignored.

The
[`hazelcast-docker`](https://github.com/hazelcast/hazelcast-docker/pull/100)
silently maps `HZ_LICENCE_KEY` to `HZ_LICENSE_KEY` to workaround this
issue, and I think therefore the behaviour is inconsistent outside of
docker.

I've added `hazelcast.enterprise.licence.key` as a
`HazelcastProperty#deprecatedName`.
This isn't quite the same behaviour - instead of using the argument
silently, instead the following is logged:
```
Don't use deprecated 'hazelcast.enterprise.licence.key' but use 'hazelcast.enterprise.license.key' instead. The former name will be removed in the next Hazelcast release.
```

This isn't strictly true (as we won't remove it), but it does make it
work with a one-line code change, and gives users a warning it's wrong
and encouragement to fix it.

After this change, we *may* want to update `hazelcast-docker` to pass
the erroneous argument through (i.e. as
`hazelcast.enterprise.licence.key`), so the error is reported there as
well.

Fixes #13161