Best-effort fix for merging metadata over WAN after merge rejection
As demonstrated in the reproducer for HZ-3392, it is possible for metadata to be incorrectly applied on WAN targets after the entry has been rejected by the supplied merge policy.

Unfortunately, due to limitations of the existing `SplitBrainMergePolicy`, there is no information passed about whether the `merge()` call modified the outcome or not (whether the incoming entry was rejected, accepted, or altered). This limitation has been documented in HZ-3397 to be addressed in HZ 6.0 as modifications would almost certainly break existing usage (such as custom merge policies).

As a short-term solution that helps improve this scenario (but does not fully resolve it), we can make an "educated guess" about whether the `merge()` call altered the outcome by comparing the returned value against the original value - if they are equal, it's likely the `merge()` did not modify the outcome.

This solution does not cover all scenarios, as it's possible for an operation to modify TTL while setting the same key/value - in this case the expiration would not be merged, resulting in the target expiration metadata always being preferred. Once the merge policy shortfalls have been addressed, this scenario can be resolved.

This commit also adds more TTL-based regression tests, including one for the future scenario outlined in the paragraph above (annotated with `@Ignore` for now and documented in HZ-3397 for removal in the future).

Fixes https://hazelcast.atlassian.net/browse/HZ-3392