Remove JUnit4 dependency management - inherit from `junit-bom` (#25696)
Hazelcast uses `junit-bom`.
`junit-bom` specifies a version for `junit-vintage-engine`.
`junit-vintage-engine` specifies a version for `junit` (i.e. JUnit4).

Hazelcast *also* specifies, via the `${junit-jupiter.version}` property,
a managed version of `junit`.

It's redundant to specify the version of `junit`, as
`junit-bom`effectively specifies the same version - but if it didn't,
we'd have compatibility issues, which is the problem `junit-bom` seeks
to avoid.

Removed `${junit.version}` property, `junit` managed dependency &
`hazelcast-archunit-rules/pom.xml` to ensure that a managed version of
`junit` is still available.