Replace `${log4j2.version}` with `log4j-bom` (#25695)
https://github.com/hazelcast/hazelcast/pull/25680 is failing because an
updated dependency uses a different version of `log4j-core` to the one
one declared in Hazelcast.

A workaround is to specify the `log4j-core` version in
`dependencyManagement` - we are already specifying a version of
`log4j-api` here.

Alternatively, we can **replace** the `log4j-api` version specification
with `log4j-bom`. This delivers the same result, as it asserts the
version of `log4j-core`, `log4j-api` etc.

Currently the version of log4j dependencies used through the project is
controlled by `${log4j2.version}`, but using `log4j-bom` removes this -
and ensures that the log4j dependencies used are all compatible with
each other.

As such I've removed the `${log4j2.version}` variable, and all it's
applications from the child poms.