Introduce ImmutableLazyEntry [HZ-2703]
Add ImmutableLazyEntry which considers it's keys and values immutable
and avoids deserialization when the key or value are not accessed. When
serializing it re-uses already serialized Data.

It is important that the key and value are immutable - any modifications
of the value will not be reflected in the serialized version.

The class is similar to
com.hazelcast.internal.journal.DeserializingEntry and
com.hazelcast.map.impl.LazyMapEntry, but provides smaller set of
interfaces and the behaviour for ser/de is different.