Lazy init of WAN user code on map container creation (#25700)
Reverts change #20419 which was added to resolve a deadlock which could
occur as a result of expensive initialization of map store context. This
was reported as a customer issue and an investigation had identified the
root cause as slow user configured WAN plugins.
This original fix introduced race conditions to the mutation of
invalidationListenerCounters which was not reported until #25327
uncovered the assertion error. The race condition is reproducible by
concurrently creating & destroying an IMap with the same name. The
problem is that the EventListenerCounter class is unable to
differentiate between old (destroyed) and new maps. Therefore this
counter must be tighter coupled to the MapContainer.

It appeared as though this was a regression, however it was found that
the test was not running with `-ea ` prior to the first reported failure
(assertions were not enabled). Running the test with the original PR
#20419 proved that the issue always existed.

In addition to reverting the original fix, this change wraps WAN
specific code from MapContainer in a new Context class and defers the
initialization of WAN user code (`DelegatingWanScheme`) until required.

Therefore when MapContainer is created to mutate event listener counters
the expensive wan context will not be eagerly initialized.
 
` MapContainer mapContainer =
mapServiceContext.getMapContainer(mapName);
`
Fixes #25327

EE PR: https://github.com/hazelcast/hazelcast-enterprise/pull/6640

Checklist:
- [X] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [X] Label `Add to Release Notes` or `Not Release Notes content` set
- [X] Request reviewers if possible