Reset delegate supplier in case of config change [HZ-3070] (#25826)
In enterprise, it is possible to re-init the WAN Context on a
configuration change; this raises the possibility that there will no
longer be a WanReplicationScheme for a Map which previously had one (on
a previous initialization). The result is that there is potential to
reference an old DelegatingWanScheme after initialization. Nullifying
the delegate on initialization ensures consistency with the original
logic.

Checklist:
- [X] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [X] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [X] Request reviewers if possible