Revert change to method references in `CacheDataSerializerHook` (#25835)
In https://github.com/hazelcast/hazelcast/pull/25785,
`CacheDataSerializerHook.createFactory()` was changed from generating an
array of `ConstructorFunctions` to `Suppliers` - and as a result
`java.lang.NoClassDefFoundError: javax/cache/CacheException` was thrown
in a test.

More specifically, from `ConstructorFunctions` declared via a lambda
function, to `Supplier` declared via a method reference. Method
references aren't just neater, they are also faster - by directly
binding (via `MetaFactory`) to the to the target method. I would
**_guess_** that this direct binding is causing an earlier resolution of
the class' dependencies, and identification that a dependency of
`provided` scope is missing, causing an exception where previously this
did not happen.

To address this, I've changed `CacheDataSerializerHook.createFactory()`
back to using a lambda, but still with a `Supplier`, and the tests now
pass without issue, and documented this oddity for the future.

Fixes https://github.com/hazelcast/hazelcast/issues/25834