Add SQL integration test
The test is checking the DataSource is really closed - it wasn't. Fixed
(TransformBatchedP wasn't closing the traverser) and made some
additional changes.

However, I am not sure if adding AutoCloseable to Traverser is a good
idea - Traversers are composed and I would expect a closeable Traverser,
used with e.g. Traverser#map to produce again a closeable Traverser.

Maybe a better solution is to record the iterator(s) created in
JdbcJoinFullScanProcessorSupplier (many) and
JdbcJoinPredicateScanProcessorSupplier (just 1) and close them in
`Processor#close` method there.