TPC connection count default to 1.
Instead of connecting to all TPC ports, by default, a client will
connect random ports of the offered TPC ports from the server.

The reason behind this is that connecting to many ports, lead to
smaller and more packets and this leads to networking overhead
and running into bottlenecks (e.g. the pps limit on AWS).

The connection count is configurable from the config and from the
System.properties and if needed, one can connect to all TPC ports
and shoot a request directly into the right core.