Introduce MergingValue#getDeserializedValue [HZ-3701]
During the RCA for an issue related to data being deserialized by
mistake during WAN replication, it was agreed that the MergingValue
class's `getValue` method is misleading in that it deserializes the
value, whereas most `getValue` methods in the codebase return the
in-memory value without deserializing it.

To decrease the likelihood of deserialization being invoked when it
shouldn't be, this commit deprecates `MergingValue#getValue` and
introduces `MergingValue#getDeserializedValue` to replace it. This
commit also improves javadocs for further clarity to developers.