[TPC] Tpc setTpcTokenToEndpoint fix (#25937)
I broke the authentication token setting logic in
https://github.com/hazelcast/hazelcast/pull/25486/ by not implementing
setTpcTokenToEndpoint in AuthMessageTask and CustomAuthMessageTask.

This PR fixes that. Also we should migrate some simple test cases of
ClientTpcTest to a new test class that'll be a Quick test, to be able to
catch regressions like this in the future. Let's discuss which tests
should be Quick test in this PR.

---------

Co-authored-by: Peter Veentjer <alarmnummer@gmail.com>