Move early-rewriting `HepProgram` initialization to static context (#26032)
We have two separate optimization phases, where we are unconditionally
rewriting some relations (e.g, converting all projects and filters to
`Calc`, etc). During this phase **for each query** we create a
`HepProgram` (set of rules) for a `HepPlanner`, which is completely
redundant: we're wasting extra allocations and CPU time for the context,
which factically is static.

This patch moves most of the unconditionally rewriting rules into a
static/per-instance context. Worth noticing, that we can't move every rule to the
static `HepProgram`, since we're checking for cyclic UDTs usage during
the unconditional rewriting phase. UNION rewrites also were moved to a
second phase: they are not required during the first phase, and moving
them to the static context preserves extra allocations.

Signed-off-by: Sasha Syrotenko <oleksandr.syrotenko@hazelcast.com>